# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: PropTech Platform API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /properties:
    get:
      operationId: listProperties
      tags: [properties]
      summary: List properties
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: project
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of properties
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Property"
    post:
      operationId: createProperty
      tags: [properties]
      summary: Create property
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePropertyBody"
      responses:
        "201":
          description: Property created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
  /properties/{id}:
    get:
      operationId: getProperty
      tags: [properties]
      summary: Get property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateProperty
      tags: [properties]
      summary: Update property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdatePropertyBody"
      responses:
        "200":
          description: Updated property
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Property"
    delete:
      operationId: deleteProperty
      tags: [properties]
      summary: Delete property
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # CONTRACTS
...
  /import/preview:
    post:
      operationId: previewImport
      tags: [import]
      summary: Preview import data
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImportPreviewBody"
      responses:
        "200":
          description: Preview result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportPreviewResponse"
  /import/commit:
    post:
      operationId: commitImport
      tags: [import]
      summary: Commit import
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommitImportBody"
      responses:
        "200":
          description: Import result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJob"
  /import/jobs/{id}:
    get:
      operationId: getImportJob
      tags: [import]
      summary: Get import job detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Import job
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJob"

  # RENTAL - TENANTS
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Property:
      type: object
      properties:
        id:
          type: integer
        companyId:
          type: ["integer", "null"]
        projectName:
          type: string
        block:
          type: ["string", "null"]
        floor:
          type: ["integer", "null"]
        unitNumber:
          type: string
        type:
          type: string
          enum: [apartment, office, commercial, parking, storage, house]
        area:
          type: ["number", "null"]
        status:
          type: string
          enum: [available, sold, reserved, rented, on_lease, archived]
        rentalStatus:
          type: ["string", "null"]
          enum: [free, rented, overdue, archived, null]
        comment:
          type: ["string", "null"]
        externalId:
          type: ["string", "null"]
        sourceType:
          type: ["string", "null"]
        syncStatus:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - projectName
        - unitNumber
        - type
        - status
        - createdAt
        - updatedAt
    CreatePropertyBody:
      type: object
      properties:
        projectName:
          type: string
        block:
          type: ["string", "null"]
        floor:
          type: ["integer", "null"]
        unitNumber:
          type: string
        type:
          type: string
          enum: [apartment, office, commercial, parking, storage, house]
        area:
          type: ["number", "null"]
        status:
          type: string
          enum: [available, sold, reserved, rented, on_lease, archived]
        comment:
          type: ["string", "null"]
        externalId:
          type: ["string", "null"]
      required:
        - projectName
        - unitNumber
        - type
        - status
    UpdatePropertyBody:
      type: object
      properties:
        projectName:
          type: string
        block:
          type: ["string", "null"]
        floor:
          type: ["integer", "null"]
        unitNumber:
          type: string
        type:
          type: string
          enum: [apartment, office, commercial, parking, storage, house]
        area:
          type: ["number", "null"]
        status:
          type: string
          enum: [available, sold, reserved, rented, on_lease, archived]
        comment:
          type: ["string", "null"]
...
    ImportJob:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [counterparties, properties, contracts]
        status:
          type: string
          enum: [pending, processing, completed, failed, partial]
        totalRows:
          type: integer
        successRows:
          type: integer
        errorRows:
          type: integer
        errors:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - type
        - status
        - totalRows
        - successRows
        - errorRows
        - createdAt
        - updatedAt
    ImportPreviewBody:
      type: object
      properties:
        type:
          type: string
          enum: [counterparties, properties, contracts]
        data:
          type: array
          items:
            type: object
      required:
        - type
        - data
    ImportPreviewResponse:
      type: object
      properties:
        totalRows:
          type: integer
        validRows:
          type: integer
        errorRows:
          type: integer
        errors:
          type: array
          items:
            $ref: "#/components/schemas/ImportRowError"
        preview:
          type: array
          items:
            type: object
      required:
        - totalRows
        - validRows
        - errorRows
        - errors
        - preview
    ImportRowError:
      type: object
      properties:
        row:
          type: integer
        field:
          type: ["string", "null"]
        message:
          type: string
      required:
        - row
        - message
    CommitImportBody:
      type: object
      properties:
        type:
          type: string
          enum: [counterparties, properties, contracts]
        data:
          type: array
          items:
            type: object
        onlyValid:
          type: boolean
      required:
        - type
        - data
...